+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.8.0] - (20.02.2025)
+
+## Changes
+
+- fix enpower grid state error
+- added encharge grid state sensor
+- added solar grid state sensor
+- bump dependencies
+- cleanup
+
 ## [9.7.6] - (18.02.2025)
 
 ## Changes